move node views into node editor folder
