+++ src/backoffice/editors/node/views/editor-view-node-edit.element.ts
@@ -1,9 +1,9 @@
 import { css, html, LitElement } from 'lit';
 import { UUITextStyles } from '@umbraco-ui/uui-css/lib';
 import { customElement, state } from 'lit/decorators.js';
-import { NodeEntity, NodeProperty } from '../../mocks/data/content.data';
-import { UmbContextConsumerMixin } from '../../core/context';
-import { UmbNodeContext } from '../editors/node/node.context';
+import { NodeEntity, NodeProperty } from '../../../../mocks/data/content.data';
+import { UmbContextConsumerMixin } from '../../../../core/context';
+import { UmbNodeContext } from '../node.context';
 import { Subscription, distinctUntilChanged } from 'rxjs';
 
 @customElement('umb-editor-view-node-edit')
